+++ R/space_filling.R
@@ -15,8 +15,7 @@
 #' `"audze_eglais"`, `"max_min_l1"`, `"max_min_l2"`, `"uniform"`,
 #' `"max_entropy"`, or `"latin_hypercube"`. A value of `"any"` will choose the
 #' first design available (in the order listed above, excluding
-#' `"latin_hypercube"`). If the design is extremely small, the function may
-#' change the type to `"latin_hypercube"` (with a warning).
+#' `"latin_hypercube"`). For a single-point design, a random grid is created.
 #' @param variogram_range A numeric value greater than zero. Larger values
 #'  reduce the likelihood of empty regions in the parameter space. Only used
 #'  for `type = "max_entropy"`.
@@ -92,8 +91,11 @@
 grid_space_filling <- function(x, ..., size = 5, type = "any", original = TRUE) {
   dots <- list(...)
   if (any(names(dots) == "levels")) {
-    rlang::warn(
-      "`levels` is not an argument to `grid_space_filling()`. Did you mean `size`?"
+    cli::cli_abort(
+      c(
+        "{.arg levels} is not an argument to {.fn grid_space_filling}.",
+        i = "Did you mean {.arg size}?"
+      )
     )
   }
   UseMethod("grid_space_filling")
@@ -192,10 +194,9 @@ make_sfd <- function(...,
   params <- map(param_quos, eval_tidy)
   p <- length(params)
 
-  if (size < p | size == 1) {
-    cli::cli_warn("Due to the small size of the grid, a Latin hypercube \\
-                  design will be used.")
-    type <- "latin_hypercube"
+  if (size == 1) {
+    res <- grid_random(params, size = size)
+    return(res)
   }
 
   if (type %in% c("any", "audze_eglais", "max_min_l1", "max_min_l2", "uniform")) {
@@ -282,7 +283,12 @@ grid_max_entropy <- function(x,
 
   dots <- list(...)
   if (any(names(dots) == "levels")) {
-    rlang::warn("`levels` is not an argument to `grid_max_entropy()`. Did you mean `size`?")
+    cli::cli_abort(
+      c(
+        "{.arg levels} is not an argument to {.fn grid_max_entropy}.",
+        i = "Did you mean {.arg size}?"
+      )
+    )
   }
   UseMethod("grid_max_entropy")
 }
@@ -409,7 +415,12 @@ grid_latin_hypercube <- function(x, ..., size = 3, original = TRUE) {
 
   dots <- list(...)
   if (any(names(dots) == "levels")) {
-    rlang::warn("`levels` is not an argument to `grid_latin_hypercube()`. Did you mean `size`?")
+    cli::cli_abort(
+      c(
+        "{.arg levels} is not an argument to {.fn grid_latin_hypercube}.",
+        i = "Did you mean {.arg size}?"
+      )
+    )
   }
   UseMethod("grid_latin_hypercube")
 }