+++ tests/testthat/test-grids.R
@@ -1,11 +1,5 @@
 
 test_that("regular grid", {
-  expect_error(
-    grid_regular(mtcars)
-  )
-  expect_error(
-    grid_regular()
-  )
   expect_snapshot(
     error = TRUE,
     grid_regular(mixture(), trees(), levels = 1:4)
@@ -13,6 +7,10 @@ test_that("regular grid", {
   expect_equal(
     nrow(grid_regular(mixture(), trees(), levels = 2)),
     4
+  )
+  expect_equal(
+    nrow(grid_regular(mixture(), trees(), levels = 2, filter = trees == 1)),
+    2
   )
   expect_equal(
     nrow(grid_regular(mixture(), trees(), levels = 2:3)),
@@ -44,15 +42,14 @@ test_that("regular grid", {
     grid_regular(list(mixture(), trees()), levels = 3),
     grid_regular(mixture(), trees(), levels = 3)
   )
+
+  expect_snapshot(
+    error = TRUE,
+    grid_regular(mixture(), trees(), levels = c(2, trees = 4))
+  )
 })
 
 test_that("random grid", {
-  expect_error(
-    grid_random(mtcars)
-  )
-  expect_error(
-    grid_random()
-  )
   expect_equal(
     nrow(grid_random(mixture(), trees(), size = 2)),
     2
@@ -63,6 +60,10 @@ test_that("random grid", {
     nrow(grid_random(prod_degree(), prune_method(), size = 50)),
     12
   )
+  expect_equal(
+    nrow(grid_random(list(mixture(), trees()), size = 2)),
+    2
+  )
 })
 
 
@@ -70,11 +71,15 @@ test_that("wrong argument name", {
   skip_if_below_r_version("3.6")
   p <- parameters(penalty(), mixture())
   set.seed(1)
-  
-  expect_snapshot(grid_space_filling(p, levels = 5, type = "latin_hypercube"))
-  expect_snapshot(grid_space_filling(p, levels = 5, type = "max_entropy"))
-  expect_snapshot(grid_random(p, levels = 5))
-  expect_snapshot(grid_regular(p, size = 5))
+
+  expect_snapshot(error = TRUE, {
+    grid_space_filling(p, levels = 5, type = "latin_hypercube")
+  })
+  expect_snapshot(error = TRUE, {
+    grid_space_filling(p, levels = 5, type = "max_entropy")
+  })
+  expect_snapshot(error = TRUE, grid_random(p, levels = 5))
+  expect_snapshot(error = TRUE, grid_regular(p, size = 5))
 })
 
 test_that("filter arg yields same results", {
@@ -97,6 +102,13 @@ test_that("filter arg yields same results", {
 test_that("new param grid from conventional data frame", {
   x <- data.frame(num_comp = 1:3)
 
-  expect_error(y <- dials:::new_param_grid(x), regexp = NA)
+  expect_no_condition(y <- dials:::new_param_grid(x))
   expect_true(tibble::is_tibble(y))
+
+  # or from a matrix?
+  expect_snapshot(
+    error = TRUE,
+    new_param_grid(as.matrix(x))
+  )
+
 })